+++ app/src/main/java/dev/tuandoan/tasktracker/data/database/TaskDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.RoomDatabase
 
 @Database(
     entities = [Task::class],
-    version = 10,
+    version = 11,
     exportSchema = true,
 )
 abstract class TaskDatabase : RoomDatabase() {